+++ R/checkArgumentsPDTMT.R
@@ -13,7 +13,7 @@
     modificationsCol, excludeUnmodifiedPeptides, keepModifications,
     iColPattern, sampleAnnot, includeOnlySamples, excludeSamples,
     minScore, minDeltaScore, minPeptides, minPSMs, masterProteinsOnly,
-    imputeMethod, assaysForExport, addHeatmaps, mergeGroups,
+    imputeMethod, assaysForExport, addAbundanceValues, addHeatmaps, mergeGroups,
     comparisons, ctrlGroup, allPairwiseComparisons, singleFit,
     subtractBaseline, baselineGroup, normMethod, spikeFeatures, stattest,
     minNbrValidValues, minlFC, samSignificance, nperm, volcanoAdjPvalThr,
@@ -156,6 +156,7 @@
     .assertScalar(x = imputeMethod, type = "character",
                   validValues = c("impSeqRob", "MinProb", "MinProbGlobal"))
     .assertVector(x = assaysForExport, type = "character", allowNULL = TRUE)
+    .assertScalar(x = addAbundanceValues, type = "logical")
     .assertScalar(x = addHeatmaps, type = "logical")
     .assertScalar(x = normMethod, type = "character",
                   validValues = c(MsCoreUtils::normalizeMethods(), "none",