+++ R/checkArgumentsSpectronaut.R
@@ -11,8 +11,8 @@
     spectronautFileType, outLevel,
     spectronautLogFile, aName, idCol, labelCol, geneIdCol, proteinIdCol,
     stringIdCol, extraFeatureCols, iColPattern, sampleAnnot, includeOnlySamples,
-    excludeSamples, minScore,
-    minPeptides, imputeMethod, assaysForExport, addHeatmaps, mergeGroups,
+    excludeSamples, minScore, minPeptides, imputeMethod, assaysForExport, 
+    addAbundanceValues, addHeatmaps, mergeGroups,
     comparisons, ctrlGroup, allPairwiseComparisons, singleFit,
     subtractBaseline, baselineGroup, normMethod, spikeFeatures, stattest,
     minNbrValidValues, minlFC, samSignificance, nperm, volcanoAdjPvalThr,
@@ -146,6 +146,7 @@
     .assertScalar(x = imputeMethod, type = "character",
                   validValues = c("impSeqRob", "MinProb", "MinProbGlobal"))
     .assertVector(x = assaysForExport, type = "character", allowNULL = TRUE)
+    .assertScalar(x = addAbundanceValues, type = "logical")
     .assertScalar(x = addHeatmaps, type = "logical")
     .assertScalar(x = normMethod, type = "character",
                   validValues = c(MsCoreUtils::normalizeMethods(), "none",