+++ R/runMaxQuantAnalysis.R
@@ -80,6 +80,8 @@
 #'     if another assay is used for the actual analysis and comparison of
 #'     groups. If set to \code{NULL} or an assay name that does not exist in
 #'     the SingleCellExperiment object, the 'main' assay will be used.
+#' @param addAbundanceValues Logical scalar defining whether abundance values
+#'     for individual samples should be added to the exported result tables.  
 #' @param addHeatmaps Logical scalar indicating whether to include heatmaps
 #'     or not. This controls both the heatmap showing the missing value
 #'     pattern in the data, as well as the summary heatmaps of the
@@ -269,7 +271,8 @@ runMaxQuantAnalysis <- function(
     iColPattern, sampleAnnot,
     includeOnlySamples = "", excludeSamples = "",
     minScore = 10, minPeptides = 2, imputeMethod = "MinProb",
-    assaysForExport = c("iBAQ", "Top3"), addHeatmaps = TRUE,
+    assaysForExport = c("iBAQ", "Top3"), 
+    addAbundanceValues = TRUE, addHeatmaps = TRUE,
     mergeGroups = list(), comparisons = list(),
     ctrlGroup = "", allPairwiseComparisons = TRUE, singleFit = TRUE,
     subtractBaseline = FALSE, baselineGroup = "", normMethod = "none",
@@ -319,7 +322,8 @@ runMaxQuantAnalysis <- function(
         includeOnlySamples = includeOnlySamples,
         excludeSamples = excludeSamples, minScore = minScore,
         minPeptides = minPeptides, imputeMethod = imputeMethod,
-        assaysForExport = assaysForExport, addHeatmaps = addHeatmaps,
+        assaysForExport = assaysForExport, 
+        addAbundanceValues = addAbundanceValues, addHeatmaps = addHeatmaps,
         mergeGroups = mergeGroups,
         comparisons = comparisons, ctrlGroup = ctrlGroup,
         allPairwiseComparisons = allPairwiseComparisons, singleFit = singleFit,
@@ -372,8 +376,9 @@ runMaxQuantAnalysis <- function(
              includeOnlySamples = includeOnlySamples,
              excludeSamples = excludeSamples, minScore = minScore,
              minPeptides = minPeptides, imputeMethod = imputeMethod,
-             assaysForExport = assaysForExport, addHeatmaps = addHeatmaps,
-             mergeGroups = mergeGroups,
+             assaysForExport = assaysForExport, 
+             addAbundanceValues = addAbundanceValues, 
+             addHeatmaps = addHeatmaps, mergeGroups = mergeGroups,
              comparisons = comparisons, ctrlGroup = ctrlGroup,
              allPairwiseComparisons = allPairwiseComparisons,
              singleFit = singleFit,