+++ tests/testthat/test-runFragPipeAnalysis.R
@@ -37,6 +37,7 @@ test_that("runFragPipeAnalysis works", {
         minPeptides = 2,
         imputeMethod = "MinProb",
         assaysForExport = NULL,
+        addAbundanceValues = TRUE, 
         addHeatmaps = TRUE,
         mergeGroups = list(),
         comparisons = list(),
@@ -298,6 +299,15 @@ test_that("runFragPipeAnalysis works", {
     expect_error(do.call(runFragPipeAnalysis, args),
                  "'assaysForExport' must be of class 'character'")
 
+    ## addAbundanceValues
+    args <- args0
+    args$addAbundanceValues <- 1
+    expect_error(do.call(runFragPipeAnalysis, args),
+                 "'addAbundanceValues' must be of class 'logical'")
+    args$addAbundanceValues <- c(TRUE, FALSE)
+    expect_error(do.call(runFragPipeAnalysis, args),
+                 "'addAbundanceValues' must have length 1")
+    
     ## addHeatmaps
     args <- args0
     args$addHeatmaps <- 1